--- module/model-postgres/test/sql.ts
@@ -1,1 +1,0 @@
-// Obsolete multi-table SQL query test suite. JSON document-store uses standard model-query tests.
